+++ betfair/src/main/java/com/smt/betfair/dto/request/Filter.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class Filter {
     private List<Integer> eventTypeIds;
-    private List<Integer> eventIds;
+    private List<Long> eventIds;
     private List<String> marketTypeCodes;
 }
